requirements.txt
google-cloud-storage
Flask
mecab-python3
Werkzeug
